requirements_llm.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
langgraph
python-dotenv
langchain-mcp-adapters
mcp[cli]
pytz